backend2/requirements.txt
fastapi
uvicorn
pydantic
librosa
numpy
openai
google-generativeai
python-multipart
python-dotenv